created _energy_cal.py for energy calibration tools
