+++ rqpy/core/__init__.py
@@ -3,3 +3,4 @@ from ._pulse import *
 from ._fitting import *
 from ._rrq import *
 from ._iv_didv_tools import *
+from ._energy_cal import *
